Refactored heavily the DisplayFragment
