+++ app/src/main/java/com/github/factotum_sdp/factotum/ui/display/client/ClientPhotoDiffCallback.kt
@@ -1,10 +1,10 @@
-package com.github.factotum_sdp.factotum.ui.display.utils
+package com.github.factotum_sdp.factotum.ui.display.client
 
 import androidx.recyclerview.widget.DiffUtil
 import com.google.firebase.storage.StorageReference
 
 // Callback to calculate the difference between two photo items
-class PhotoDiffCallback : DiffUtil.ItemCallback<StorageReference>() {
+class ClientPhotoDiffCallback : DiffUtil.ItemCallback<StorageReference>() {
 
     override fun areItemsTheSame(oldItem: StorageReference, newItem: StorageReference): Boolean {
         return oldItem.path == newItem.path
